----------------------------------------------------------------------
Committing in new version of releasing-hol, which correctly builds the
non-standard documentation for the n_bit library.  The non-standard
part is that the library is named "n_bit" but the documentation is
named "word.{dvi,pdf}".

The documentation is renamed to "<version>-n_bit.{dvi,pdf}" for the
exported, public Kananaskis-3.

Anthony, please review this and let me know if you would like some
corrections.

Modified Files:
	releasing-hol
----------------------------------------------------------------------