Added "let" to the input language. Scott Owens noticed that
    let v = M in N

can be mapped to

   Seq (Par (\x.x) (\args. M)) (\(args,v). N)

so we don't have to add anything extra to get computation sharing via
lets. This should also transfer over to the hardware front-end I hope.
Now we can write

   Round ((y,z),(k0,k1,k2,k3),s)  =
      let s' = s + DELTA in
      let y' = ShiftXor(z, s', k0, k1)
      in
       ((y', z + ShiftXor(y', s', k2, k3)), (k0,k1,k2,k3), s')

for TEA, which looks much nicer and produces 1/3 smaller code because we
no longer duplicate the expression abbreviated by y'.