Fixing typos in the documentation (included as a comment in sexp.sml).
The method of converting a full ACL2 name pkg::sym to a HOL-friendly
name is as follows (see definition of create_hol_name below).

 1. First see if the ACL2 name already has a HOL name, and if so
    return it.

 2. Convert sym to a HOL name, say hol_sym, using the ML function
    acl2_name_to_hol_name (this function is described below).

 3. Check to see if sym_name is already used. If not it is the result.

 4. If hol_sym is in use, then use acl2_name_to_hol_name to convert pkg to
    hol_pkg and then see if (hol_pkg ^ "_" ^ hol_sym) is used, and if
    not return it. If it is used an error report is printed and an
    exception raised.

 5. Record any names generated by step 1-4 above in acl2_names.

The function acl2_name_to_hol_name converts a string s to a
HOL-friendly name as follows.

 1. Convert s to a list of characters and convert all letters
    to lower-case.

 2. Replace "-" by "_".

 3. Replace special characters by strings of letter characters
    using acl2_to_hol_char_map (defined below). Add an underscore
    "_" to separate special characters from all other characters.

We think these simple rules should be sufficient in practice, and if
they are not it's better to reconsider carefully what to do rather
than to have ugly looking names generated by more complicated rules
(this might change if the sufficiency assumption is badly wrong).