Replace the temporary acl2_defs_theory by hol_defaxiomsTheory.
Initially hol_defaxiomsScript.sml started as a list of desired
definitions (and -- later in the file -- theorems) imported from
ACL2. The imported desired stuff is in the comments.

For example, after the boilerplate, hol_defaxiomsScript.sml starts
with:

(*
     [oracles: DEFUN ACL2::IFF, DISK_THM] [axioms: ] []
     |- iff p q = itel [(p,andl [q; t]); (q,nil)] t,
*)

This comment contains a printout of a theorem created by trusting ACL2
(see added tag, which indicates that DEFUN is trusted and showing the
ACL2 name). The long-term goal is to eliminate the need to trust ACL2 when
defining the core ACL2 theory. Thus we have the pure HOL definition:

val iff_def =
 acl2Define "ACL2::IFF"
  `iff p q = itel [(p,andl [q; t]); (q,nil)] t`;

So far we have only made HOL definitions of those functions
immediately needed (e.g. by James' translate tool).

Sometimes, for historical reasons, we define a function slightly
different from the definition in the comment (e.g. see ACL2::ZP) and
then prove that the definition made implements the one in the comment.

Eventually desired recursive definitions will be needed. So far
acl2Define doesn't support this well, so we have the interim
implementation of

(*
     [oracles: DEFUN ACL2::TRUE-LISTP, DISK_THM] [axioms: ] []
     |- true_listp x = ite (consp x) (true_listp (cdr x)) (eq x nil),
*)

implemented as:

val (true_listp_def,true_listp_ind) =
 Defn.tprove
  (Hol_defn
    "true_listp"
    `true_listp x = ite (consp x) (true_listp (cdr x)) (eq x nil)`,
   WF_REL_TAC `measure sexp_size`
    THEN RW_TAC arith_ss [sexp_size_cdr]);

but eventually we need the constant to be named "ACL2::TRUE-LISTP",
and "true_lisp" is just overloaded onto that.

Once all the definitions have been made, we hope to prove the ACL2
axioms, thereby establishing that we have a model of the ACL2
theory. This may be a lot of work though.