The ARM model can now be run using one of two programs "run" (MLton) or "mosrun.exe" (Mosml).
For example:

./run -E -pc 0x20 -cycles 100 file.s

will load the code from "file.s" on-top of a minimal exception handler.
A run is then performed for 100 cycles, with an initial PC value of 0x20.
Two output files are generated: "run.in" and "run.out", which give the
initial and final states.

Options can be listed with ./run --help .