: ---------------------------------------------------------------------- : Enter Log.  Lines beginning with `CVS:' are removed automatically : : Committing in additions to pattern-matching for constants, and : to fix bound variables in nested cases. : : Modified Files: : 	hol98/src/parse/Literal.sig hol98/src/tfl/src/Functional.sig : 	hol98/src/tfl/src/Functional.sml : ----------------------------------------------------------------------
The purpose is to support natural (:num), character, and string
constants within the patterns that are accepted and processed by
Functional.mk_functional.  This is used in Define, and probably
elsewhere.  It also attempts to fix the rather poor choices of
mk_functional for the names of bound variables to better match
the names given in the input to the function.

This version at least permits the HOL system to build, but I cannot
guarantee that it is perfect.

I thought that it might also fix the following parsing anomaly.
Try typing in

``(checkPrefix :
     ('a -> bool) -> 'a list # 'a list -> 'a list option) (aR :'a -> bool)
     ((subl :'a list),(supl :'a list)) =
   case (subl,supl) of
      (([] :'a list),(v3 :'a list)) -> SOME supl
   || ((subhd :'a)::(subtl :'a list),([] :'a list)) ->
        (NONE :'a list option)
   || (subhd::subtl,(suphd :'a)::(suptl :'a list)) ->
        (if subhd = suphd then
           checkPrefix aR (subtl,suptl)
         else
           (NONE :'a list option))``;

I get

> val it =
    ``checkPrefix aR (subl,supl) =
      case (subl,supl) of
         ([],v1) -> SOME supl
      || (v2::v3,[]) -> NONE
      || (v2::v3,v4::v5) ->
           (if v2 = v4 then checkPrefix aR (v3,v5) else NONE)`` :
  term

which is a rather unfortunate renaming!  I'd be happy to try to
repeat the fixing procedure that I implemented for mk_functional
for the parsing of nested cases, but I don't know where this is.
Michael or Konrad, do you have an idea?

Peter