---------------------------------------------------------------------- Enter Log.  Lines beginning with `CVS:' are removed automatically
Committing in revisions to support XEmacs as well as Emacs.
The tools/hol98-mode.src file now should work for both editors.
There are two new functions early in the file that should be used
in the future for either detecting if a region is selected, or
turning on the selection.  This has been tested, and I think it's
solid, but if anyone has problems, please contact me.

Modified Files:
	configure.sml hol98-mode.src
----------------------------------------------------------------------