----------------------------------------------------------------------
Committing in general expressions in patterns.
This extends literals in patterns to general expressions, to permit
literals of all types to be used in "case" constructs.
This is rough code, but because of the demand, I thought I'd get
this out there now so people can try it, and find errors if possible.
I'll be polishing this up soon.

Modified Files:
	hol98/src/bool/Pmatch.sml hol98/src/bool/TypeBasePure.sml
	hol98/src/tfl/src/Functional.sml
----------------------------------------------------------------------