Occurrences of "iN" have been replaced by "N".
The system specification has been generalised to NEXT_ARM_SYS.
This makes the model "coprocessor and MMU ready".  This development
has been put in systemTheory.

The function NEXT_ARM_MEM is now a specialisation of NEXT_ARM_SYS
to the case of a simple memory model with no coprocessors.

The upshot for users is that the state space

arm_mem_state

is replaced by

('a0,'a1,'a2,'a3,'a4,'a5,'a6,'a7,
 'a8,'a9,a10,'a11,'a12,'a13,'a14) arm_sys_state

which now contains the coprocessor registers.  The type variables
correspond with the register length for each coprocessor e.g. could
be ``:256``, ``:8`` or simply ``:32``.   In arm_rulesTheory, the
coprocessor registers are never modified because each coprocessor
is always "absent".