1. The coprocessor interface has been re-implemented.  (In light of the fact that it didn't fit in with the way the ARM VFP coprocessor works.)
The upshot is that the state space has changed again (this time for
the simpler).  The new state space is:

'a arm_sys_state =
   <| registers : registers;
      psrs      : psrs;
      memory    : mem;
      undefined : bool;
      cp_state  : 'a
   |>

Here, ``:'a`` represents that state of *all* the coprocessors.  The
theorems in arm_rulesTheory should be the same (modulo the changed
type).

The behaviour of the coprocessors is specified with an object of type:

'a coproc =
  <| absent : bool -> word32 -> bool;
     f_cdp  : 'a -> bool -> word32 -> 'a;
     f_mrc  : 'a -> bool -> word32 -> word32;
     f_mcr  : 'a -> bool -> word32 -> word32 -> 'a;
     f_stc  : 'a -> bool -> word32 -> word32 option list;
     f_ldc  : 'a -> bool -> word32 -> word32 list -> 'a
  |>

The function ``ADD_COPROC : 'a coproc -> 'b coproc -> ('a # 'b) coproc``
can be used to combine coprocessor specifications.

Some changes have been made to enable evaluation with coprocessors.
There are two new functions: arm_evalLib.evaluate_cp and eval_cp.
All evaluations are now wrt the function STATE_ARM_MMU, which takes
a coprocessor as its first argument.  systemTheory.STATE_ARM_MEM is
``STATE_ARM_MMU NO_CP``, where NO_CP corresponds with all coprocessors
being absent.

2. The specification of "data transfer" instructions has been changed
in armTheory (i.e. LDR_STR, LDM_STM and SWP).  The return type is now
a "sum" type (instead of being a record) and the data input is now an
option type.  This has enabled ``ARB`` parameters to be eliminated,
which avoids exceptions in the EmitML code.