Post processing for word parsing is supported with the function:
wordsLib.guess_word_lengths : term -> term

This was requested by Joe Hurd.

When possible, word lengths are guessed for the extract and
concatenate functions.  That is,  ``(a >< b) w`` is given
type ``: (a + 1 - b) word`` and ``(a: a word) @@ (b: b word)``
is given type ``: (a + b) word``.

For example:

- val _ = Parse.post_process_term := wordsLib.guess_word_lengths;
- ``(3 >< 0) a @@ (7 >< 4) a @@ (16 >< 8) a``;
<<HOL message: inventing new type variable names: 'a, 'b, 'c, 'd, 'e, 'f>>
<<HOL message: assigning word length(s): 'a <- 4, 'b <- 13, 'c <- 17, 'e <- 4 and 'f <- 9>>
> val it = ``(3 >< 0) a @@ (7 >< 4) a @@ (16 >< 8) a`` : term

The assignment message is controlled by the trace variable
"notify word length guesses".