The compiler from TFL (in a special form) to ARM machine code has been updated to allow proper procedure calls. Files with major changes:
  examples/mc-logic/arm_compilerLib.sml
  examples/mc-logic/arm_compiler_demoScript.sml

The updates enable proper procedure calls with context 
saving/restoring. Now the compiler can compile a function 
with one of the following options:

 - InLineCode
 - SimpleProcedure
 - PushProcedure (regs,depth)

The InLineCode option makes any call to the compiled 
function just the function body itself. The 
SimpleProcedure and PushProcedure both result in proper 
procedure calls. The SimpleProcedure can only be used 
(and is recommended) for procedures that do not call 
other procedures and do not require temporary space on 
the stack. A call to a procedure compiled with the 
PushProcedure (regs,depth) option will preserve the 
value of registers named in the list regs; it will also 
make `depth` number of temporary spaces available on the 
stack. I've added documentation with examples to the 
tutorial in arm_compiler_demoScript.sml