Revised higher order quotient package to give better error messages in case the user input is incomplete.  In particular, a message is printed describing the missing element in case any of the following are not supplied by the user:
1. quotient type extension theorem
2. quotient simplification theorems
3. respectfulness theorem for each constant being lifted
4. polymorphic respectfulness theorem for each polymorphic constant in a theorem being lifted, and
5. polymorphic preservation theorem for each polymorphic constant in a theorem being lifted.

For cases 2 and 3 above, the form of the needed theorem is printed.  For the other cases, it is not, but this might be added later.  In any case the name of the type or constant in question is printed, focusing the user's search for the cause of the quotient package's failure.

Also, in all cases above, the name of the input field where the new theorem is to be added is specified in the error message.