Improved the type parsing and the rank inference.  Removed the distinction between tuple type application and type application, making them both one kind of application that associates to the left.  The appearance of a tuple of types means the same curried application as before, where (ty1,ty2,ty3) opr = ty3 (ty2 (ty1 opr)).
The removal of the distinction between tuple type application and type application means that

``:('c,'d) ('a,'b) 'z``;

will not parse now.  The intended type would be entered as

``:('c,'d) (('a,'b) 'z)``;

The big advantage of unifying the two kinds of type applications is that any kind of postfix type operator can now appear after a parenthesized type, such as type and rank casts:

val ty11 = ``:('a:<=2) : ar 5``;

The new term forms for type abstractions and combinations have not been added to the term parser yet.