Thanks to Larry Paulson for pointing out that tail-recursive functions don't need termination proofs to be admitted into HOL's logic. As a result, these files have been simplified.
Termination proofs (which are still needed by our methodology) can now
be integrated naturally into induction proofs (as shown in the first
example of decompiler_demoTheory, which used to be 35 lines long, but
is now only 14 lines long. The main lemma behind this advance is
TAILREC_PRE_THM in tailrecTheory.