Mike and I have prepared these examples for a paper submitted to SBMF 2008. These examples presents how programs such as this
  a := 0;
  while not (n = 0) do
    a := a + 1;
    n := n - 2
  end

can automatically be translated into recursive functions, such as

  d (a,n) = 
    let a = 0 in d_1 (a,n)

  d_1 (a,n) = 
    if n = 0 then (a,n) else
      let a = a + 1 in 
      let n = n - 2 in 
        d_1 (a,n)

In the process the automation proves a total-correctness specification
relating the extracted function (`d` above) with the original program: 

  |- d_pre (a,n) ==>
     SEP_TOTAL_SPEC 
       (VAR "a" a * VAR "n" n)
       (Seq (Assign "a" (Const 0))
         (While (Not (Equal (Var "n") (Const 0)))
           (Seq (Assign "a" (Plus (Var "a") (Const 1)))
                (Assign "n" (Sub (Var "n") (Const 2))))))
       (let (a,n) = d (a,n) in (VAR "a" a * VAR "n" n))

The precondition `d_pre` is defined as follows. It states (in an
induction-friendly way) that `d` terminates.

  d_pre (a,n) = 
    let a = 0 in d_1_pre (a,n)

  d_1_pre (a,n) = 
    ~(n = 0) ==>
      let a = a + 1 in 
      let n = n - 2 in 
        d_1_pre (a,n)

The presented translation into recursive functions is a simplification
of what the machine-code decompiler in HOL/examples/mc-logic does with
x86, ARM and PowerPC machine code.


Ack. Thanks to Larry Paulson for pointing out that the tail-recursive
function `WHILE` can be defined without a termination proof.

  WHILE P g x = if P x then WHILE P g (g x) else x