Add some support for nice Unicode constants.  Start up hol, type   UTF8_Printing.all_printing() and then   pred_setTheory.IN_UNION
If you run HOL inside an xterm, you may need to have LANG environment
variables set.  If you are using HOL inside Emacs, you will probably
need: 
  (setq default-buffer-file-coding-system 'utf-8)
  (setq process-coding-system-alist '((".*" . utf-8)))
in your .emacs file.

Please experiment with making this nicer.  (It may also be nice to
provide users with a way of backing out of this :-)