Implement a scheme whereby a simpset can be limited to apply no more than a limited number of rewrites/conversions/d.p. invocations.  The limit is stored in a simpset.  By default all simpsets are unlimited. To limit a simpset, use   limit n simpset and it will then stop simplifying after n actions.
There is a border case: if a conditional rewrite requires exactly as
many rewrites to solve its side-conditions as it is limited to, it
will still go through with the rewrite, thereby exceeding the limit by
one (n side conditions, plus the final simplification)

The design is there to give a (better) guarantee of termination than
simpLib currently provides.  For this reason, I didn't go with the
otherwise appealing idea of just letting side-condition discharge
ignore the limit entirely.