Set up some code to provide monad syntax in the style of Haskell.  The calls to temp_overload_on at the bottom of the monadsyntax.sml file illustrate how specific constants implementing the monad primitives can be used in the syntax. (Actually, it's just the monad_bind and monad_unitbind that are required; I just prefer "return" to "UNIT".)
This syntax allows
   do v <- action ; u <- action2 v ; return (u,v) od
to parse to the appropriate thing and to print appropriately too.  In
HOL, we can't make just one overload and have it work for all monad
types, but I guess in HOL-Omega, there might be a monad_bind constant
that would stand in for all possible monads.  (Or maybe I'm confused
about HOL-Omega...)

Soon to add: support for a parallel operator allowing

  do u <- action1 !! v <- action2 ; foobar u v od

where !! implements some action combiner such that the above is
actually equivalent to

  do (u,v) <- (action1 !! action2) ; foobar u v od

which is in turn equivalent to

  monad_bind (action1 !! action2) (\(u,v). foobar u v)