Fixed many bugs, in particular in the higher order matching of both types and terms, revising those algorithms substantially.
Extended the functorScript.sml example in examples/HolOmega to a fairly nice tutorial example.  There are two versions: functor1Script.sml has tactics that allow the user to step through the proof at a low level, seeing the transformation of the goal under the tactics, while functorScript.sml emphasizes the use of the simplifier to shorten the scripts considerably.

There is also a new example, aopScript.sml, which is just begun and very small at present.