Update the handling of binders so that different concrete syntax forms can map to the same abstract syntax form.  This makes binders more similar to other forms.  For example, the if-then-else and =>-| concrete forms might both map to the underlying name "COND".  Now, with Unicode enabled, the Unicode universal quantifier as well as "!" can both map to the name "!".  (Internally, there is a preferred flag set, so that the pretty-printer knows which should be printed when it comes to the term "!".)
This couldn't be done before, so for example the Unicode universal
quantifier mapped to itself, and this name was then overloaded to be
an alias for "!".  This use of the overloading level for Unicode led
to some problems, and was inconsistent with the approach taken for the
other symbols.  

There are some other tidy-ups now that binders are more like other
forms.