Fixed many bugs, in particular in the higher order matching of types, so that it succeeds in matching more cases.  Still, it is possible to construct a type which cannot be matched to some instances of the type.
Chiefly, extended the monadScript.sml example in examples/HolOmega to a tutorial example.  Monads may be characterized either by the operations "unit" and "bind" with three laws, or by the operations "unit", "map", and "join" with seven laws; the two specifications are shown to be equivalent.  Also, "map" is shown to be a functor, and "unit" and "join" are shown to be natural transformations.  Several examples of monads are given, including a state monad.