Concrete syntax can be such a pain.  Anthony has pointed out that we need to be able to parse both
  r with fld := value

and

  r with <| ...  ; fld := subrcd with subfld := subvalue |>

Both of these seem as if the forms above should be accepted. This
requires "with" and ":=" (and their unpopular friend "updated_by"), to
all be at the same level and right-associative.  I've put them at 460,
making them looser than the loosest value operator, which is
left-associated ++ (at 480).  This ensures that

  fld := x ++ y

parses correctly.