This was causing hundreds of empty temporary files to be generated during the build on PolyML.  This quick fix
 	       val tmp          = FileSys.tmpName()
-	       val cnfname      = if isSome fname then (valOf fname) else tmp^".cnf";
+	       val cnfname      = if isSome fname then (valOf fname) else tmp;

isn't ideal, but fixed the bulk of the problem for now.  On PolyML, the call to
tmpName() actually creates the file (as per the ML basis standars), and by
adding the suffix, the original file is never removed.  There are still obvious
potential cases where the temporary file is created but not used or removed,
but I don't immediately see how to fix them (or indeed what this code does).