Made substantial changes to the monad example in examples/HolOmega.  There are now three independent definitions of monads, one of which is based on category theory notions from functorScript.sml.  All three are proved equivalent.  The category theory version uses new operators defined in functorScript for the composition of functors, the composition of functors with natural transformations on the left or on the right, and the vertical composition of natural transformations.  These are notions and proofs which could not have been accomplished without the new features of HOL-Omega.
Several bugs and changes to the underlying structure of types and terms were corrected, including the higher order matching of types, overloading of term constants, and the higher order matching of terms of the form ``var [: type1, ..., typen :] arg1 ... argm``.   In particular, higher order matching of types now throws an exception if the algorithm terminates with a substitution which does not in fact transform the pattern into the target.  A bug was found in the core, in the definition of eta-conversion.  A couple of incompleteness bugs in type beta reduction were found and corrected.  The theorems EQ_TY_EXT and TY_FUN_EQ_THM were improved for kind variables and generality.

The HOL-Omega system is coming along.  Although not bullet-proof, and still showing a high bug rate, it is beginning to become a useful tool, as evidenced by the monad example in examples/HolOmega.

Somehow the current version of HOL-Omega had drifted apart from the main branch of Kananaskis-5.  Some of these changes in the src/parse directory have been manually installed, but I am not sure I found all of them, and further checks are warranted.