Modify lexing so that "mixed" tokens are possible if a user puts them into the grammar.  For example, one might do
  set_fixity "-->a" (Infix(NONASSOC, 450))

This then causes ``p -->a q`` to be parsed as an infix application of
-->a to p and q.  If you wrote ``p --> a q``, the -->a token would not
be seen.  Without -->a being in the grammar, it will be lexed as
before, a symbolic token (-->) followed by an alphanumeric one (a).
(Note also that things like --a--> where the mixing up happens in the
middle of the token, are just as possible.)

This change also allows traditional alphanumeric identifiers to
combine with Unicode subscripts.  Using LaTeX notation, a_1 is now
possible.

The system builds with selftest level 1, so I believe I have preserved
backwards compatibility.  (I have also added a test to
src/bool/selftest.sml to check for _1 working as a case expression
pattern properly.)