Revised the examples to have the natural transformation predicate take the natural transformation as its first argument (instead of last).
Fixed the type unification of the parser to not do higher-order type inference in the presence of unbound unification variables, since they could later be bound to the newly introduced bound variables, causing confusion.  Calculating this restriction for safe unification required manual dereferencing of unification variables bound, not through the reference, but via the accumulated environment.  This new restriction on type inference had some ripple effects on the examples.