Add some useful functions   get_first : ('a -> 'b option) -> 'a list -> 'b option   bool_compare : bool * bool -> order   flip_cmp : ('a * 'a -> order) -> ('a * 'a -> order)
The get_first function is useful if you want to find the guy that
can be transformed successfully by some function, and you want the
result of the transformation, not the input to it.  (Otherwise, you'd
use List.find (isSome o f).)

The get_first name is taken from Isabelle's standard library (I've
used it multiple times under different names throughout HOL).

bool_compare should be in the Basis Library but isn't.