Revised the monad examples further, making the types of the MMAP, JOIN, and BIND operators to produce values of universal types.  In the state monad example, the "state_bind" operator is now overloaded with an infix binary operator ">>=", which displays as infix even though it has a universal type!
These revisions exposed some bugs, one of which was serious, concerning the higher order matching of types.  The corrected version is pretty solid now.  However, it is important to remember that with the more expressive types and terms of HOL-Omega, matching may not complete even if there is a substitution which unifies the pattern with the target.  This is in contrast to normal higher order logic, where matching is guaranteed to complete if a substitution exists.  Even in that case, however, higher order matching is not guaranteed to complete, and even if it does, the currently implemented algorithms may not produce the right substitution.