Changes so that the weakened rewriting of lambda-terms now happens with respect to beta-equality (==), rather than beta-reduction (-b->*).  This gives us access to more rewrites.  In particular, we get things like   church (SUC n) @@ z @@ f == f @@ (church n @@ z @@ f) which is a nice recursive characterisation of what church numerals do, and saves us from having to do too many ghastly things with FUNPOW. With the above equation, neither side reduces to the other (they have a common reduct of course), and so it wasn't a usable rewrite when using -b->*.
In addition, to the above, I set up a natrec combinator, which would
get typed 
  : 'a -> (num -> 'a -> 'a) -> num -> 'a
if we were in a typed world.  This combinator gets access to the value
of the recursive argument, making predecessor easier to define.  

Also, give an equation like the one above for church for the cDB
function.

Many proofs are shorter.  Some are slower, but the shortness of
expression won me over.