Handle the discrepancy in the behaviour between the way CHOOSE_THEN and GEN_TAC choose variants by making them do the same thing (Konrad earlier fixed GEN_TAC to use prim_variant, but CHOOSE_THEN was using variant).
Decided that avoiding constant names in the current grammar was
probably a good idea, so open up the internal gen_variant function in
Term.sml, and feed it with the new predicate "is_constname" from
Parse.  is_constname is true if a string is the name of a "constant"
as the parser sees things.  In particular, if you do
  overload_on ("foo", ``SUC``)
then "foo" (as well as "SUC") will be considered a const's name.

Hidden constants are not const names in this sense, so C doesn't get
primed in either of
  STRIP_TAC ([], ``(?C. P C) ==> Q``)
or
  STRIP_TAC ([], ``!C. P C ==> Q``)
even with combinTheory as an ancestor.