Fix for bug reported by Tom Ridge.  Change to ind_types.sml includes some reformatting/re-indentation of the source code around the change, so it looks more significant than it is.  It turns out that the problem was in the behaviour of the sorting function (which seems to be the place where the HOL Light and HOL4 implementations diverge: in their underlying libraries!).
The code gets the list of type arguments to constructors that mention
types about to be defined underneath other operators.  For example, in

  'a list = NIL | CONS of 'a => 'a list

there are none.  In complicated cases like Tom's, there are lots.  He
had

  com = skip | seq of bool # com # com | if of bool # num # com # com
      | while of (num # num) # bool # com

Then the list of types is

  [bool # (com # com), bool # (num # (com # com)), (num # num) # (bool # com)]

Each of these is dealt with in turn, and it is important that they be
done in order with bigger types being done first.  The constructors in
each argument are broken down one level at a time, so the
implementation would eventually get itself into a state where the list
of recursive nested types was something like

  [com # com, bool # com, num # (com # com)]

Unfortunately, in this situation, the sorting function did the wrong
thing and didn't see that

   num # (com # com)

should be done before (com # com).  The ordering function passed to
the sort function was the flip of sub-type inclusion, with
incomparable types being treated as EQUAL.  In the situation above,
num # (com # com) didn't move to the front of the list.  This sort of
thing is a risk with a sorting comparison that doesn't respect
equality.  (I.e.,  num # (com # com) < com # com, but it isn't less
than bool # com, which is, however, 'equal' to com # com.)

I see that HOL Light has a quicksort as its sort function, whereas our
Listsort.sort (from the Moscow ML library) is a mergesort.  Perhaps
quicksort does provide some sort of useful guarantee in this
situation.   Anyway, the fix is easy enough: just sort by (flip of)
type size.