Prove some theorems about normal order reduction to weak-head normal forms (abstractions, or variables in the head position of a series of applications).  Basically, if something can normal reduce to such a thing, then it can weak-head reduce to something that shares the key aspects of the shape.  Either, it can weak-head reduce to an abstraction, or it can weak-head reduce to a series of applications, where the head is the variable that is in that position at the end of the normal order reduction.
Along the way, invent a notation for
  FOLDL (@@) t ts
where t is a term, and ts a list of terms.  The notation is (t @* ts)
in ASCII.  This is a "series of applications" with t in the head
position.  Prove a bunch of theorems about these forms.   All this
stuff goes into basics/appFOLDLTheory.