A proper fix for what r6851 attempted to fix.  The problem was demonstrated by ideaMultScript in examples/Crypto/IDEA.  This loaded its ancestor theory ExtendedEuclid before it loaded the ind_types module.  The problem is that ExtendedEuclid makes "P" a constant, and ind_types attempts to prove a lemma internally where P has to be a variable.  Now, it also sets up its parser to be with respect to a grammar not contaminated by ExtendedEuclid, so its statement of the goal is fine.
The goal is

  !P t. (!x. (x = t) ==> P x) ==> $? P

Unfortunately, when the tactic that comes to solve this goal calls

  REPEAT GEN_TAC

the code in GEN_TAC spots the "P" and checks the *global* grammar for
clashes with known constants.  P is such a beast, and so the goal
after the GEN_TAC is

  (!x. (x = t) ==> P' x) ==> $? P'

A later tactic that mentions P (which is still parsed correctly as a
variable called "P") then causes the tactic to derail and the proof to
fail.

It's possible to make the tactic work by using X_GEN_TAC (but not
Q.X_GEN_TAC), but it's all terribly ugly, and so I decided to prove
the theorem in boolTheory, so that the *three* places where the same
code was being duplicated could pick up the theorem without reproving
it.  I also removed the code duplication by making the one version in
Prim_rec the source for the two other users.

Now examples/Crypto/IDEA builds properly.

This fix will have to be merged into

  branches/release-development/kananaskis-5

This does suggest, incidentally, that Konrad's solution to parsing
protection, involving setting the global grammar reference at the head
of library files, is pragmatically better, for all that futzing around
with references is ugly, and requires the end of the file to set the
global reference back to what it was (yergh).  I will have to give up
on my "clean" solution, which rebinds/redefines the Parse structure at
the head of library files.