This is a major update of HOL-Omega, which is now extended to the experimental kernel and to PolyML.  Those who prefer to use the experimental kernel can now do
build -expk

in the installation process, and all the new functionality of HOL-Omega will be present in the resulting theorem prover.

The standard kernel may be built with either MoscowML or with PolyML.  The experimental kernel currently requires PolyML, as it does not build correctly with MoscowML (just as the normal HOL4 experimental kernel does not build with MoscowML).  Hopefully this will be rectified in the future.

There are some minor incompatibilities between the standard kernel and the experimental kernel.  Those differences are retained in the HOL-Omega versions, as each HOL-Omega kernel is a backwards-compatible extension of the corresponding HOL4 kernel.  Anyone who has been using a particular HOL4 kernel should see no differences by moving to the HOL-Omega version of that kernel.